rename protocol dialogs to panels
The SGE protocol calls its second window kind a "dialog": a fixed layout
of widgets with sizes and coordinates, as opposed to a stream of text.
That word means a blocking modal box both to users and in Compose, which
is the opposite of what these are, and it collided with the app's real
dialogs.

Rename the concept to "panel" throughout, giving a three-level
vocabulary: stream / panel / widget (one element inside a panel).

DialogObject -> PanelObject, DialogState -> PanelState,
ComposeDialogState -> ComposePanelState, WindowType.DIALOG ->
WindowType.PANEL, DialogWindowData -> PanelWindowData,
WraythDialog*Event -> WraythPanel*Event, getOrCreateDialog ->
getOrCreatePanel, DialogContent -> PanelContent, and friends.

The wire tag strings ("dialogdata", "opendialog") keep the protocol's
word, as do the real modal dialogs, androidx's DialogWindow /
rememberDialogState, and FileKit's dialogs package. No string literal
changed, so this is behavior-neutral.
