+++ compose/src/commonMain/kotlin/warlockfe/warlock3/compose/ui/window/ComposePanelState.kt
@@ -3,17 +3,17 @@ package warlockfe.warlock3.compose.ui.window
 import kotlinx.collections.immutable.toPersistentList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import warlockfe.warlock3.core.client.DialogObject
-import warlockfe.warlock3.core.window.DialogState
+import warlockfe.warlock3.core.client.PanelObject
+import warlockfe.warlock3.core.window.PanelState
 
-class ComposeDialogState(
+class ComposePanelState(
     override val id: String,
-) : DialogState {
-    private val cachedList = mutableListOf<DialogObject>()
-    private val _objects = MutableStateFlow(emptyList<DialogObject>())
+) : PanelState {
+    private val cachedList = mutableListOf<PanelObject>()
+    private val _objects = MutableStateFlow(emptyList<PanelObject>())
     val objects = _objects.asStateFlow()
 
-    override suspend fun setObject(value: DialogObject) {
+    override suspend fun setObject(value: PanelObject) {
         cachedList.removeAll { it.id == value.id }
         cachedList.add(value)
     }
