+++ compose/src/commonMain/kotlin/warlockfe/warlock3/compose/ui/window/PanelButton.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.input.pointer.pointerHoverIcon
 import androidx.compose.ui.unit.Dp
 
 @Composable
-fun DialogButton(
+fun PanelButton(
     onClick: () -> Unit,
     shape: Shape,
     background: (isHovered: Boolean, isPressed: Boolean) -> Brush,
