--- core/src/commonMain/kotlin/warlockfe/warlock3/core/window/DialogState.kt
@@ -1,13 +1,0 @@
-package warlockfe.warlock3.core.window
-
-import warlockfe.warlock3.core.client.DialogObject
-
-interface DialogState {
-    val id: String
-
-    suspend fun setObject(value: DialogObject)
-
-    suspend fun clear()
-
-    suspend fun updateState()
-}
