+++ wrayth/src/commonMain/kotlin/warlockfe/warlock3/wrayth/protocol/elements/OpenPanelHandler.kt
@@ -2,16 +2,16 @@ package warlockfe.warlock3.wrayth.protocol.elements
 
 import warlockfe.warlock3.wrayth.protocol.BaseElementListener
 import warlockfe.warlock3.wrayth.protocol.StartElement
-import warlockfe.warlock3.wrayth.protocol.WraythDialogWindowEvent
 import warlockfe.warlock3.wrayth.protocol.WraythEvent
-import warlockfe.warlock3.wrayth.util.WraythDialogWindow
+import warlockfe.warlock3.wrayth.protocol.WraythPanelWindowEvent
+import warlockfe.warlock3.wrayth.util.WraythPanelWindow
 
-class OpenDialogHandler : BaseElementListener() {
+class OpenPanelHandler : BaseElementListener() {
     override fun startElement(element: StartElement): WraythEvent? {
         // We receive a location here that we're ignoring to have a bit more control over where things are placed
         val id = element.attributes["id"] ?: return null
-        return WraythDialogWindowEvent(
-            WraythDialogWindow(
+        return WraythPanelWindowEvent(
+            WraythPanelWindow(
                 id = id,
                 title = element.attributes["title"] ?: id,
                 type = element.attributes["type"],
