+++ wrayth/src/commonMain/kotlin/warlockfe/warlock3/wrayth/protocol/elements/PanelDataHandler.kt
@@ -2,15 +2,15 @@ package warlockfe.warlock3.wrayth.protocol.elements
 
 import warlockfe.warlock3.wrayth.protocol.BaseElementListener
 import warlockfe.warlock3.wrayth.protocol.StartElement
-import warlockfe.warlock3.wrayth.protocol.WraythDialogDataEvent
 import warlockfe.warlock3.wrayth.protocol.WraythEvent
+import warlockfe.warlock3.wrayth.protocol.WraythPanelDataEvent
 
-class DialogDataHandler : BaseElementListener() {
+class PanelDataHandler : BaseElementListener() {
     override fun startElement(element: StartElement): WraythEvent =
-        WraythDialogDataEvent(
+        WraythPanelDataEvent(
             id = element.attributes["id"],
             clear = element.attributes["clear"]?.startsWith(prefix = "t", ignoreCase = true) == true,
         )
 
-    override fun endElement(): WraythEvent = WraythDialogDataEvent(null, clear = false)
+    override fun endElement(): WraythEvent = WraythPanelDataEvent(null, clear = false)
 }
