+++ wrayth/src/commonMain/kotlin/warlockfe/warlock3/wrayth/protocol/elements/ProgressBarHandler.kt
@@ -1,18 +1,18 @@
 package warlockfe.warlock3.wrayth.protocol.elements
 
-import warlockfe.warlock3.core.client.DialogObject
+import warlockfe.warlock3.core.client.PanelObject
 import warlockfe.warlock3.core.client.Percentage
 import warlockfe.warlock3.wrayth.protocol.BaseElementListener
 import warlockfe.warlock3.wrayth.protocol.StartElement
-import warlockfe.warlock3.wrayth.protocol.WraythDialogObjectEvent
 import warlockfe.warlock3.wrayth.protocol.WraythEvent
+import warlockfe.warlock3.wrayth.protocol.WraythPanelObjectEvent
 import warlockfe.warlock3.wrayth.util.parseDistance
 
 class ProgressBarHandler : BaseElementListener() {
     override fun startElement(element: StartElement): WraythEvent? {
         val id = element.attributes["id"] ?: return null
-        return WraythDialogObjectEvent(
-            DialogObject.ProgressBar(
+        return WraythPanelObjectEvent(
+            PanelObject.ProgressBar(
                 id = id,
                 value = element.attributes["value"]?.let { Percentage.fromString(it) } ?: Percentage(0),
                 text = element.attributes["text"],
