+++ wrayth/src/commonMain/kotlin/warlockfe/warlock3/wrayth/util/WraythPanelWindow.kt
@@ -1,6 +1,6 @@
 package warlockfe.warlock3.wrayth.util
 
-data class WraythDialogWindow(
+data class WraythPanelWindow(
     val id: String,
     val title: String,
     val location: String?,
